refactor(step1): app에서 security를 사용하는 부분을 app.security로 이동
